backend/requirements.txt
alembic==0.9.2
aniso8601==1.2.1
bcrypt==3.1.3
cffi==1.10.0
click==6.7
coverage==4.4.1
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.0.4
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
pkg-resources==0.0.0
pycparser==2.17
python-dateutil==2.6.0
python-decouple==3.0
python-editor==1.0.3
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.10
Werkzeug==0.12.2
